refactor: extract da light client initialization
